routing configured. write page added and display an information
Need to fix the problem where we try to put navigation button with
topics and link to the page with writes which related to the current
topic. There are problem with reloading write page, because an
information cannot reload from database. Need to find out how to reload
controller when an information that controller uses not change in the
WritesController.js